+++ branches/2.0/doubango/tinySIGCOMP/src/tcomp_udvm.bytecopy.c
@@ -36,9 +36,9 @@
 /**RFC3320-Setction_8.4.  Byte copying
 From UDVM to UDVM
 */
-int tcomp_udvm_bytecopy_self(tcomp_udvm_t *udvm, uint16_t *destination, uint16_t source, uint16_t tsk_size_tocopy)
+int tcomp_udvm_bytecopy_self(tcomp_udvm_t *udvm, uint32_t *destination, uint32_t source, uint32_t tsk_size_tocopy)
 {
-	uint16_t byte_copy_left, byte_copy_right;
+	uint32_t byte_copy_left, byte_copy_right;
 	//if (*destination == TCOMP_UDVM_GET_SIZE() || source == TCOMP_UDVM_GET_SIZE())
 	if (*destination >= TCOMP_UDVM_GET_SIZE() || source >= TCOMP_UDVM_GET_SIZE())
 	{
@@ -69,9 +69,9 @@ int tcomp_udvm_bytecopy_self(tcomp_udvm_t *udvm, uint16_t *destination, uint16_t
 /**RFC3320-Setction_8.4.  Byte copying
 From EXTERNAL to UDVM
 */
-int tcomp_udvm_bytecopy_to(tcomp_udvm_t *udvm, uint16_t destination, const uint8_t* source, uint16_t tsk_size_tocopy)
+int tcomp_udvm_bytecopy_to(tcomp_udvm_t *udvm, uint32_t destination, const uint8_t* source, uint32_t tsk_size_tocopy)
 {
-	uint16_t byte_copy_left, byte_copy_right;
+	uint32_t byte_copy_left, byte_copy_right;
 
 	if(destination == TCOMP_UDVM_GET_SIZE())
 	{
@@ -101,9 +101,9 @@ int tcomp_udvm_bytecopy_to(tcomp_udvm_t *udvm, uint16_t destination, const uint8
 /**RFC3320-Setction_8.4.  Byte copying
 From UDVM to EXTERNAL
 */
-int tcomp_udvm_bytecopy_from(tcomp_udvm_t *udvm, uint8_t* destination, uint16_t source, uint16_t tsk_size_tocopy)
+int tcomp_udvm_bytecopy_from(tcomp_udvm_t *udvm, uint8_t* destination, uint32_t source, uint32_t tsk_size_tocopy)
 {
-	uint16_t byte_copy_left, byte_copy_right;
+	uint32_t byte_copy_left, byte_copy_right;
 
 	if(source == TCOMP_UDVM_GET_SIZE())
 	{
